# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Pocket Trader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 2 */
  username: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  username: string;
  credits: number;
  botActive: boolean;
  autoConfirm: boolean;
  isAdmin: boolean;
  banned: boolean;
  /** @nullable */
  pocketOptionId: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface AdminUser {
  id: number;
  email: string;
  username: string;
  credits: number;
  botActive: boolean;
  autoConfirm: boolean;
  isAdmin: boolean;
  banned: boolean;
  /** @nullable */
  pocketOptionId: string | null;
  totalTrades: number;
  createdAt: string;
}

export interface AdminUserUpdate {
  banned?: boolean;
  isAdmin?: boolean;
}

export interface CreditsAdjust {
  credits: number;
}

export interface BotStatusUpdate {
  botActive?: boolean;
  autoConfirm?: boolean;
}

export interface PocketOptionUpdate {
  pocketOptionId: string;
}

export type SignalDirection = typeof SignalDirection[keyof typeof SignalDirection];


export const SignalDirection = {
  BUY: 'BUY',
  SELL: 'SELL',
  HOLD: 'HOLD',
} as const;

export interface Signal {
  direction: SignalDirection;
  confidence: number;
  rsi: number;
  ema9: number;
  ema21: number;
  asset: string;
  timeframe: string;
}

export type TradeInputDirection = typeof TradeInputDirection[keyof typeof TradeInputDirection];


export const TradeInputDirection = {
  BUY: 'BUY',
  SELL: 'SELL',
} as const;

export interface TradeInput {
  asset: string;
  timeframe: string;
  amount: number;
  direction: TradeInputDirection;
  confidence: number;
}

export type TradeDirection = typeof TradeDirection[keyof typeof TradeDirection];


export const TradeDirection = {
  BUY: 'BUY',
  SELL: 'SELL',
} as const;

export type TradeResult = typeof TradeResult[keyof typeof TradeResult];


export const TradeResult = {
  win: 'win',
  loss: 'loss',
} as const;

export interface Trade {
  id: number;
  userId: number;
  asset: string;
  timeframe: string;
  amount: number;
  direction: TradeDirection;
  result: TradeResult;
  createdAt: string;
}

export interface TradeStats {
  totalTrades: number;
  wins: number;
  losses: number;
  winRate: number;
  creditsSpent: number;
}

export type TopupInputPackage = typeof TopupInputPackage[keyof typeof TopupInputPackage];


export const TopupInputPackage = {
  NUMBER_100: '100',
  NUMBER_500: '500',
  NUMBER_1000: '1000',
} as const;

export type TopupInputCurrency = typeof TopupInputCurrency[keyof typeof TopupInputCurrency];


export const TopupInputCurrency = {
  ton: 'ton',
  usdt: 'usdt',
} as const;

export interface TopupInput {
  package: TopupInputPackage;
  txHash: string;
  currency: TopupInputCurrency;
}

export type TopupStatus = typeof TopupStatus[keyof typeof TopupStatus];


export const TopupStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Topup {
  id: number;
  userId: number;
  package: string;
  credits: number;
  usdAmount: number;
  tonAmount: number;
  txHash: string;
  currency: string;
  status: TopupStatus;
  createdAt: string;
}

export type TopupWithUserStatus = typeof TopupWithUserStatus[keyof typeof TopupWithUserStatus];


export const TopupWithUserStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface TopupWithUser {
  id: number;
  userId: number;
  username: string;
  email: string;
  package: string;
  credits: number;
  usdAmount: number;
  tonAmount: number;
  txHash: string;
  currency: string;
  status: TopupWithUserStatus;
  createdAt: string;
}

export type GetSignalParams = {
asset: string;
timeframe: string;
};