# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Pocket Trader API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: users
  - name: signals
  - name: trades
  - name: topups
  - name: admin

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me/bot:
    patch:
      operationId: updateBotStatus
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me/pocket-option:
    patch:
      operationId: updatePocketOption
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PocketOptionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /signals:
    get:
      operationId: getSignal
      tags: [signals]
      parameters:
        - name: asset
          in: query
          required: true
          schema:
            type: string
        - name: timeframe
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: executeTrade
      tags: [trades]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trades/stats:
    get:
      operationId: getTradeStats
      tags: [trades]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /topups:
    get:
      operationId: listTopups
      tags: [topups]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topup"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTopup
      tags: [topups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topup"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users/{id}/credits:
    patch:
      operationId: adminAdjustCredits
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditsAdjust"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/topups:
    get:
      operationId: adminListTopups
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopupWithUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/topups/{id}/approve:
    post:
      operationId: approveTopup
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topup"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/topups/{id}/reject:
    post:
      operationId: rejectTopup
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topup"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    RegisterInput:
      type: object
      required: [email, password, username]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        username:
          type: string
          minLength: 2

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, email, username, credits, botActive, autoConfirm, isAdmin, banned, pocketOptionId, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        credits:
          type: integer
        botActive:
          type: boolean
        autoConfirm:
          type: boolean
        isAdmin:
          type: boolean
        banned:
          type: boolean
        pocketOptionId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    AdminUser:
      type: object
      required: [id, email, username, credits, botActive, autoConfirm, isAdmin, banned, pocketOptionId, totalTrades, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        credits:
          type: integer
        botActive:
          type: boolean
        autoConfirm:
          type: boolean
        isAdmin:
          type: boolean
        banned:
          type: boolean
        pocketOptionId:
          type: ["string", "null"]
        totalTrades:
          type: integer
        createdAt:
          type: string
          format: date-time

    AdminUserUpdate:
      type: object
      properties:
        banned:
          type: boolean
        isAdmin:
          type: boolean

    CreditsAdjust:
      type: object
      required: [credits]
      properties:
        credits:
          type: integer

    BotStatusUpdate:
      type: object
      properties:
        botActive:
          type: boolean
        autoConfirm:
          type: boolean

    PocketOptionUpdate:
      type: object
      required: [pocketOptionId]
      properties:
        pocketOptionId:
          type: string

    Signal:
      type: object
      required: [direction, confidence, rsi, ema9, ema21, asset, timeframe]
      properties:
        direction:
          type: string
          enum: [BUY, SELL, HOLD]
        confidence:
          type: number
        rsi:
          type: number
        ema9:
          type: number
        ema21:
          type: number
        asset:
          type: string
        timeframe:
          type: string

    TradeInput:
      type: object
      required: [asset, timeframe, amount, direction, confidence]
      properties:
        asset:
          type: string
        timeframe:
          type: string
        amount:
          type: number
        direction:
          type: string
          enum: [BUY, SELL]
        confidence:
          type: number

    Trade:
      type: object
      required: [id, userId, asset, timeframe, amount, direction, result, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        asset:
          type: string
        timeframe:
          type: string
        amount:
          type: number
        direction:
          type: string
          enum: [BUY, SELL]
        result:
          type: string
          enum: [win, loss]
        createdAt:
          type: string
          format: date-time

    TradeStats:
      type: object
      required: [totalTrades, wins, losses, winRate, creditsSpent]
      properties:
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
        creditsSpent:
          type: integer

    TopupInput:
      type: object
      required: [package, txHash, currency]
      properties:
        package:
          type: string
          enum: ["100", "500", "1000"]
        txHash:
          type: string
        currency:
          type: string
          enum: [ton, usdt]

    Topup:
      type: object
      required: [id, userId, package, credits, usdAmount, tonAmount, txHash, currency, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        package:
          type: string
        credits:
          type: integer
        usdAmount:
          type: number
        tonAmount:
          type: number
        txHash:
          type: string
        currency:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time

    TopupWithUser:
      type: object
      required: [id, userId, username, email, package, credits, usdAmount, tonAmount, txHash, currency, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        email:
          type: string
        package:
          type: string
        credits:
          type: integer
        usdAmount:
          type: number
        tonAmount:
          type: number
        txHash:
          type: string
        currency:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time