# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Pocket Trader API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  email: string;
  username: string;
  credits: number;
  botActive: boolean;
  autoConfirm: boolean;
  isAdmin: boolean;
  banned: boolean;
  /** @nullable */
  pocketOptionId: string | null;
  totalTrades: number;
  createdAt: Date;
}