# lib/api-zod/src/generated/types/adminUserUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Pocket Trader API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUserUpdate {
  banned?: boolean;
  isAdmin?: boolean;
}